+++ src/yui-throttle/HISTORY.md
@@ -4,7 +4,7 @@ YUI Throttle Change History
 @VERSION@
 ------
 
-* Throttle no longer changes the value of |this| inside the throttled function.
+* Throttle no longer changes the value of `this` inside the throttled function.
 
 3.10.0
 ------